Update README.md TASK 1: SOFTWARE SOLUTION
This course introduces the concepts of software engineering to students who have completed the core courses in programming and project management. The principles build on previously acquired concepts, switching the emphasis from programming simple routines, to engineering robust and scalable software solutions. This course does not cover programming, but provides an overview of software engineering processes, and their challenging nature focusing on the need for a disciplined approach to software engineering. A generic process framework provides the groundwork for formal process models. Prescriptive process models such as the Waterfall Model and Agile Development are included. An introduction to the elements and phases of software engineering is included which explores requirements engineering, design concepts, and software quality.